Inherit certain properties from ports connected to patch ports.
This is a bit too hard-edged at present, but does the right thing when building
a patch from scratch at least.  Something needs to be done about removing the
properties, but this is hard for outputs since the arcs aren't keyed that way.
The main problem here is that Ardour barfs on plugins with MIDI input for audio
tracks, so if you *ever* connect the control input to a MIDI anything, it will
stick and not work in an Ardour audio track.  Maybe it should be just
implemented for inputs as a stop gap...